refactor: move Ai Agent types to core
