+++ packages/botonic-plugin-flow-builder/src/types.ts
@@ -1,5 +1,7 @@
 import {
+  AiAgentArgs,
   BotContext,
+  GuardrailRule,
   InferenceResponse,
   KnowledgeBasesResponse,
   PluginPreRequest,
@@ -64,18 +66,6 @@ export type AiAgentFunction<
   aiAgentArgs: AiAgentArgs
 ) => Promise<InferenceResponse>
 
-export interface GuardrailRule {
-  name: string
-  description: string
-}
-
-export interface AiAgentArgs {
-  name: string
-  instructions: string
-  activeTools?: { name: string }[]
-  inputGuardrailRules?: GuardrailRule[]
-  sourceIds?: string[]
-}
 export type ContentFilter<
   TPlugins extends ResolvedPlugins = ResolvedPlugins,
   TExtraData = any,
@@ -121,3 +111,5 @@ export interface RatingSubmittedInfo extends HtRatingButton {
   possibleOptions: string[]
   possibleValues: number[]
 }
+
+export { AiAgentArgs, GuardrailRule }
